# hw12_13_14_15_16_calendar/api/go/api_calendar_events.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Swagger Calendar - OpenAPI 3.0
 *
 * This is a sample Calendar Server based on the OpenAPI 3.0 specification.
 *
 * API version: 1.0.12
 */

package api

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// CalendarEventsAPIController binds http requests to an api service and writes the service results to the http response
type CalendarEventsAPIController struct {
	service      CalendarEventsAPIServicer
	errorHandler ErrorHandler
}

// CalendarEventsAPIOption for how the controller is set up.
type CalendarEventsAPIOption func(*CalendarEventsAPIController)

// WithCalendarEventsAPIErrorHandler inject ErrorHandler into controller
func WithCalendarEventsAPIErrorHandler(h ErrorHandler) CalendarEventsAPIOption {
	return func(c *CalendarEventsAPIController) {
		c.errorHandler = h
	}
}

// NewCalendarEventsAPIController creates a default api controller
func NewCalendarEventsAPIController(s CalendarEventsAPIServicer, opts ...CalendarEventsAPIOption) *CalendarEventsAPIController {
	controller := &CalendarEventsAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the CalendarEventsAPIController
func (c *CalendarEventsAPIController) Routes() Routes {
	return Routes{
		"GetCalendarEvents": Route{
			"GetCalendarEvents",
			strings.ToUpper("Get"),
			"/calendar",
			c.GetCalendarEvents,
		},
		"AddCalendarEvent": Route{
			"AddCalendarEvent",
			strings.ToUpper("Post"),
			"/calendar",
			c.AddCalendarEvent,
		},
		"UpdateCalendarEventById": Route{
			"UpdateCalendarEventById",
			strings.ToUpper("Put"),
			"/calendar/{CalendarEventId}",
			c.UpdateCalendarEventById,
		},
		"DeleteCalendarEventById": Route{
			"DeleteCalendarEventById",
			strings.ToUpper("Delete"),
			"/calendar/{CalendarEventId}",
			c.DeleteCalendarEventById,
		},
	}
}

// OrderedRoutes returns all the api routes in a deterministic order for the CalendarEventsAPIController
func (c *CalendarEventsAPIController) OrderedRoutes() []Route {
	return []Route{
		Route{
			"GetCalendarEvents",
			strings.ToUpper("Get"),
			"/calendar",
			c.GetCalendarEvents,
		},
		Route{
			"AddCalendarEvent",
			strings.ToUpper("Post"),
			"/calendar",
			c.AddCalendarEvent,
		},
		Route{
			"UpdateCalendarEventById",
			strings.ToUpper("Put"),
			"/calendar/{CalendarEventId}",
			c.UpdateCalendarEventById,
		},
		Route{
			"DeleteCalendarEventById",
			strings.ToUpper("Delete"),
			"/calendar/{CalendarEventId}",
			c.DeleteCalendarEventById,
		},
	}
}

// GetCalendarEvents - Get all calendar events
func (c *CalendarEventsAPIController) GetCalendarEvents(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.GetCalendarEvents(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// AddCalendarEvent - Add a calendar event to calendar.
func (c *CalendarEventsAPIController) AddCalendarEvent(w http.ResponseWriter, r *http.Request) {
	var eventParam Event
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&eventParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertEventRequired(eventParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertEventConstraints(eventParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.AddCalendarEvent(r.Context(), eventParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// UpdateCalendarEventById - Update Calendar Event by Id.
func (c *CalendarEventsAPIController) UpdateCalendarEventById(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	calendarEventIdParam := params["CalendarEventId"]
	if calendarEventIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"CalendarEventId"}, nil)
		return
	}
	var eventParam Event
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&eventParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertEventRequired(eventParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertEventConstraints(eventParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.UpdateCalendarEventById(r.Context(), calendarEventIdParam, eventParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// DeleteCalendarEventById - Delete Calendar Event by id.
func (c *CalendarEventsAPIController) DeleteCalendarEventById(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	calendarEventIdParam := params["CalendarEventId"]
	if calendarEventIdParam == "" {
		c.errorHandler(w, r, &RequiredError{"CalendarEventId"}, nil)
		return
	}
	result, err := c.service.DeleteCalendarEventById(r.Context(), calendarEventIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}