# hw12_13_14_15_16_calendar/api/go/model_error.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Swagger Calendar - OpenAPI 3.0
 *
 * This is a sample Calendar Server based on the OpenAPI 3.0 specification.
 *
 * API version: 1.0.12
 */

package api

type Error struct {
	Code string `json:"code"`

	Message string `json:"message"`
}

// AssertErrorRequired checks if the required fields are not zero-ed
func AssertErrorRequired(obj Error) error {
	elements := map[string]interface{}{
		"code":    obj.Code,
		"message": obj.Message,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertErrorConstraints checks if the values respects the defined constraints
func AssertErrorConstraints(obj Error) error {
	return nil
}