# hw12_13_14_15_16_calendar/api/go/model_event.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Swagger Calendar - OpenAPI 3.0
 *
 * This is a sample Calendar Server based on the OpenAPI 3.0 specification.
 *
 * API version: 1.0.12
 */

package api

import (
	"time"
)

type Event struct {
	ID string `json:"ID,omitempty"`

	Title string `json:"Title,omitempty"`

	DateStart time.Time `json:"DateStart,omitempty"`

	DateEnd time.Time `json:"DateEnd,omitempty"`

	Description string `json:"Description,omitempty"`

	UserId int32 `json:"UserId,omitempty"`

	EventNotifyTime int32 `json:"EventNotifyTime,omitempty"`
}

// AssertEventRequired checks if the required fields are not zero-ed
func AssertEventRequired(obj Event) error {
	return nil
}

// AssertEventConstraints checks if the values respects the defined constraints
func AssertEventConstraints(obj Event) error {
	return nil
}